Added procedure parser

+++ src/components/controllers/mechanical_controller/mechanical_controller.c
@@ -272,7 +272,7 @@ mechanical_controller_status_t solenoid_close(valve_instance_t valve) {
 }
 
 mechanical_controller_status_t motor_set_speed(
-    int16_t speed, stepper_motor_instances_t motor) {
+    int32_t speed, stepper_motor_instances_t motor) {
   if (!controller_state.initialized) {
     MACKI_LOG_ERROR(TAG, "Mechanical controller not initialized");
     return MECHANICAL_CONTROLLER_NOT_INITIALIZED;
