+++ src/components/controllers/mechanical_controller/mechanical_controller.h
@@ -92,5 +92,5 @@ mechanical_controller_status_t solenoid_close(valve_instance_t valve);
  * @param motor The motor to set the speed of.
  * @return True if the speed was set successfully, false otherwise.
  */
-mechanical_controller_status_t motor_set_speed(int16_t speed,
+mechanical_controller_status_t motor_set_speed(int32_t speed,
                                                stepper_motor_instances_t motor);
